feat!: simplify the logging interface
Breaking the API some more - having a variadic function as the callback is not friendly
to language bindings.  Also remove logging to a file.  I will also soon get rid of
E_INFOCONT as it is not friendly to external logging APIs.  For now we still
go to stderr but in the near future it will go through Python or JavaScript logging.