refactor: nearly complete...
This code (which I am ashamed to have written) is much too
complicated for what it actually does, the complication
seems to come from the need to "push" and "pull" at the
same time.  It should be possible to do a simpler
implementation, in theory all we need is a single buffer
of size frame_size + frame_shift.